+++ backend/src/baserow/contrib/database/config.py
@@ -47,7 +47,7 @@ class DatabaseConfig(AppConfig):
         from .fields.field_types import (
             TextFieldType, LongTextFieldType, URLFieldType, NumberFieldType,
             BooleanFieldType, DateFieldType, LinkRowFieldType, EmailFieldType,
-            FileFieldType, SingleSelectFieldType
+            FileFieldType, SingleSelectFieldType, PhoneNumberFieldType
         )
         field_type_registry.register(TextFieldType())
         field_type_registry.register(LongTextFieldType())
@@ -59,6 +59,7 @@ class DatabaseConfig(AppConfig):
         field_type_registry.register(LinkRowFieldType())
         field_type_registry.register(FileFieldType())
         field_type_registry.register(SingleSelectFieldType())
+        field_type_registry.register(PhoneNumberFieldType())
 
         from .fields.field_converters import LinkRowFieldConverter, FileFieldConverter
         field_converter_registry.register(LinkRowFieldConverter())