+++ backend/src/baserow/contrib/database/fields/models.py
@@ -304,3 +304,7 @@ class FileField(Field):
 
 class SingleSelectField(Field):
     pass
+
+
+class PhoneNumberField(Field):
+    pass