+++ backend/src/baserow/contrib/database/views/view_filters.py
@@ -11,7 +11,7 @@ from pytz import timezone
 from baserow.contrib.database.fields.field_types import (
     TextFieldType, LongTextFieldType, URLFieldType, NumberFieldType, DateFieldType,
     LinkRowFieldType, BooleanFieldType, EmailFieldType, FileFieldType,
-    SingleSelectFieldType
+    SingleSelectFieldType, PhoneNumberFieldType
 )
 from .registries import ViewFilterType
 from baserow.contrib.database.fields.field_filters import contains_filter, \
@@ -38,7 +38,8 @@ class EqualViewFilterType(ViewFilterType):
         URLFieldType.type,
         NumberFieldType.type,
         BooleanFieldType.type,
-        EmailFieldType.type
+        EmailFieldType.type,
+        PhoneNumberFieldType.type
     ]
 
     def get_filter(self, field_name, value, model_field, field):
@@ -89,7 +90,8 @@ class ContainsViewFilterType(ViewFilterType):
         TextFieldType.type,
         LongTextFieldType.type,
         URLFieldType.type,
-        EmailFieldType.type
+        EmailFieldType.type,
+        PhoneNumberFieldType.type
     ]
 
     def get_filter(self, *args):
@@ -285,7 +287,8 @@ class EmptyViewFilterType(ViewFilterType):
         LinkRowFieldType.type,
         EmailFieldType.type,
         FileFieldType.type,
-        SingleSelectFieldType.type
+        SingleSelectFieldType.type,
+        PhoneNumberFieldType.type
     ]
 
     def get_filter(self, field_name, value, model_field, field):