+++ backend/tests/conftest.py
@@ -18,6 +18,26 @@ def api_client():
     return APIClient()
 
 
+def pytest_addoption(parser):
+    parser.addoption(
+        "--runslow", action="store_true", default=False, help="run slow tests"
+    )
+
+
+def pytest_configure(config):
+    config.addinivalue_line("markers", "slow: mark test as slow to run")
+
+
+def pytest_collection_modifyitems(config, items):
+    if config.getoption("--runslow"):
+        # --runslow given in cli: do not skip slow tests
+        return
+    skip_slow = pytest.mark.skip(reason="need --runslow option to run")
+    for item in items:
+        if "slow" in item.keywords:
+            item.add_marker(skip_slow)
+
+
 def run_non_transactional_raw_sql(sqls, dbinfo):
     conn = psycopg2.connect(host=dbinfo['HOST'], user=dbinfo['USER'],
                             password=dbinfo['PASSWORD'],