+++ web-frontend/modules/database/components/view/grid/fields/FunctionalGridViewFieldPhoneNumber.vue
@@ -1,0 +1,7 @@
+<template functional>
+  <div ref="cell" class="grid-view__cell">
+    <div class="grid-field-text">
+      {{ props.value }}
+    </div>
+  </div>
+</template>