+++ web-frontend/modules/database/fieldTypes.js
@@ -1,7 +1,11 @@
 import moment from 'moment'
 import BigNumber from 'bignumber.js'
 
-import { isValidURL, isValidEmail } from '@baserow/modules/core/utils/string'
+import {
+  isValidURL,
+  isValidEmail,
+  isSimplePhoneNumber,
+} from '@baserow/modules/core/utils/string'
 import { Registerable } from '@baserow/modules/core/registry'
 
 import FieldNumberSubForm from '@baserow/modules/database/components/field/FieldNumberSubForm'
@@ -20,6 +24,7 @@ import GridViewFieldBoolean from '@baserow/modules/database/components/view/grid
 import GridViewFieldDate from '@baserow/modules/database/components/view/grid/fields/GridViewFieldDate'
 import GridViewFieldFile from '@baserow/modules/database/components/view/grid/fields/GridViewFieldFile'
 import GridViewFieldSingleSelect from '@baserow/modules/database/components/view/grid/fields/GridViewFieldSingleSelect'
+import GridViewFieldPhoneNumber from '@baserow/modules/database/components/view/grid/fields/GridViewFieldPhoneNumber'
 
 import FunctionalGridViewFieldText from '@baserow/modules/database/components/view/grid/fields/FunctionalGridViewFieldText'
 import FunctionalGridViewFieldLongText from '@baserow/modules/database/components/view/grid/fields/FunctionalGridViewFieldLongText'
@@ -29,6 +34,7 @@ import FunctionalGridViewFieldBoolean from '@baserow/modules/database/components
 import FunctionalGridViewFieldDate from '@baserow/modules/database/components/view/grid/fields/FunctionalGridViewFieldDate'
 import FunctionalGridViewFieldFile from '@baserow/modules/database/components/view/grid/fields/FunctionalGridViewFieldFile'
 import FunctionalGridViewFieldSingleSelect from '@baserow/modules/database/components/view/grid/fields/FunctionalGridViewFieldSingleSelect'
+import FunctionalGridViewFieldPhoneNumber from '@baserow/modules/database/components/view/grid/fields/FunctionalGridViewFieldPhoneNumber'
 
 import RowEditFieldText from '@baserow/modules/database/components/row/RowEditFieldText'
 import RowEditFieldLongText from '@baserow/modules/database/components/row/RowEditFieldLongText'
@@ -40,6 +46,7 @@ import RowEditFieldBoolean from '@baserow/modules/database/components/row/RowEdi
 import RowEditFieldDate from '@baserow/modules/database/components/row/RowEditFieldDate'
 import RowEditFieldFile from '@baserow/modules/database/components/row/RowEditFieldFile'
 import RowEditFieldSingleSelect from '@baserow/modules/database/components/row/RowEditFieldSingleSelect'
+import RowEditFieldPhoneNumber from '@baserow/modules/database/components/row/RowEditFieldPhoneNumber'
 
 import { trueString } from '@baserow/modules/database/utils/constants'
 import {
@@ -1127,5 +1134,67 @@ export class SingleSelectFieldType extends FieldType {
       value: 'Option',
       color: 'light-blue',
     }
+  }
+}
+
+export class PhoneNumberFieldType extends FieldType {
+  static getType() {
+    return 'phone_number'
+  }
+
+  getIconClass() {
+    return 'phone'
+  }
+
+  getName() {
+    return 'Phone Number'
+  }
+
+  getGridViewFieldComponent() {
+    return GridViewFieldPhoneNumber
+  }
+
+  getFunctionalGridViewFieldComponent() {
+    return FunctionalGridViewFieldPhoneNumber
+  }
+
+  getRowEditFieldComponent() {
+    return RowEditFieldPhoneNumber
+  }
+
+  prepareValueForPaste(field, clipboardData) {
+    const value = clipboardData.getData('text')
+    return isSimplePhoneNumber(value) ? value : ''
+  }
+
+  getSort(name, order) {
+    return (a, b) => {
+      const stringA = a[name] === null ? '' : '' + a[name]
+      const stringB = b[name] === null ? '' : '' + b[name]
+
+      return order === 'ASC'
+        ? stringA.localeCompare(stringB)
+        : stringB.localeCompare(stringA)
+    }
+  }
+
+  getSortIndicator() {
+    return ['text', '0', '9']
+  }
+
+  getDocsDataType(field) {
+    return 'string'
+  }
+
+  getDocsDescription(field) {
+    return (
+      'Accepts a phone number which has a maximum length of 100 characters' +
+      ' consisting solely of digits, spaces and the following characters: ' +
+      'Nx,._+*()#=;/- .'
+    )
+  }
+
+  getDocsRequestExample(field) {
+    return '+1-541-754-3010'
   }
 }