+++ web-frontend/modules/database/plugin.js
@@ -11,6 +11,7 @@ import {
   DateFieldType,
   FileFieldType,
   SingleSelectFieldType,
+  PhoneNumberFieldType,
 } from '@baserow/modules/database/fieldTypes'
 import {
   EqualViewFilterType,
@@ -73,6 +74,7 @@ export default ({ store, app }) => {
   app.$registry.register('field', new EmailFieldType())
   app.$registry.register('field', new FileFieldType())
   app.$registry.register('field', new SingleSelectFieldType())
+  app.$registry.register('field', new PhoneNumberFieldType())
   app.$registry.register('importer', new CSVImporterType())
   app.$registry.register('importer', new PasteImporterType())
   app.$registry.register('settings', new APITokenSettingsType())