+++ changelog.md
@@ -21,6 +21,7 @@
 * Fixed SSRF bug in the file upload by URL by blocking urls to the private network.
 * Fixed bug where an invalid date could be converted to 0001-01-01.
 * The list_database_table_rows search query parameter now searches all possible field types.
+* Add Phone Number field.
 
 ## Released (2021-03-01)